refactor - extract ChartSection component
